backend/requirements.txt
   fastapi
   uvicorn
   pydantic
   psycopg2-binary  # For PostgreSQL connection